requirements.txt
asgiref==3.8.1
Django==5.2.3
django-ckeditor==6.7.3
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-modeltranslation==0.19.15
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.1
sqlparse==0.5.3